// src/api/dtos/team.ts
export type TeamDTO = {
    team: {
        id: number;
        name: string;
        country: string;
        founded: number;
        national: boolean;
        logo: string;
    };
    venue: {
        id: number;
        name: string;
        address: string;
        city: string;
        capacity: number;
        surface: string;
        image: string;
    };
};